requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==2.0.15
pydantic==1.10.8
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
email-validator==2.0.0
openai==0.27.7
python-dotenv==1.0.0
psycopg2-binary==2.9.6
redis==4.5.5
celery==5.2.7
